feat: reorganise pages
